Refactor NotaDTO package and update config files
Moved NotaDTO from dto/notaDTO to dto package and updated all relevant imports. Cleaned up application.properties by moving dev-specific settings to application-dev.properties. Updated .iml to include resources folder and added new dictionary words for spellcheck.
