+++ .idea/dictionaries/project.xml
@@ -18,6 +18,7 @@
       <w>associata</w>
       <w>associato</w>
       <w>associazioni</w>
+      <w>attivo</w>
       <w>attuale</w>
       <w>bidirezionale</w>
       <w>bidirezionali</w>
@@ -32,6 +33,7 @@
       <w>collegamenti</w>
       <w>collegamento</w>
       <w>completa</w>
+      <w>configurazione</w>
       <w>contenuto</w>
       <w>controlla</w>
       <w>controllo</w>
@@ -51,6 +53,7 @@
       <w>dell’utente</w>
       <w>deve</w>
       <w>direttamente</w>
+      <w>disattivato</w>
       <w>disegno</w>
       <w>diversi</w>
       <w>duplicata</w>
@@ -68,12 +71,14 @@
       <w>filtro</w>
       <w>finalmente</w>
       <w>forniti</w>
+      <w>gestione</w>
       <w>gestiti</w>
       <w>gestito</w>
       <w>gruppo</w>
       <w>idealmente</w>
       <w>ignora</w>
       <w>implementare</w>
+      <w>impostazioni</w>
       <w>irreversibile</w>
       <w>l'aggiornamento</w>
       <w>l'email</w>
@@ -91,6 +96,7 @@
       <w>mappiamo</w>
       <w>metodo</w>
       <w>migliorare</w>
+      <w>minimale</w>
       <w>modifica</w>
       <w>modificare</w>
       <w>nascondere</w>
@@ -107,6 +113,8 @@
       <w>prendi</w>
       <w>presente</w>
       <w>previsti</w>
+      <w>profili</w>
+      <w>profilo</w>
       <w>proprietario</w>
       <w>prova</w>
       <w>pulita</w>
