+++ Backend/src/main/resources/application-dev.properties
@@ -1,0 +1,7 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/MemoFlowDB
+spring.datasource.username=postgres
+spring.datasource.password=Adriano5495
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
