+++ Backend/src/main/resources/application.properties
@@ -1,32 +1,18 @@
 spring.application.name=MemoFlow
 
-#Gestione ENUM
+# Gestione ENUM
 spring.jpa.properties.hibernate.type.preferred_enum_type=string
 
-#Configurazione JTA
+# Configurazione JTA
 spring.jpa.properties.hibernate.transaction.jta.platform=org.hibernate.service.jta.platform.internal.JBossAppServerJtaPlatform
 
-#Debug
+# Debug
 logging.level.org.springframework.security=DEBUG
 logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 
-#SWAGGER
+# Swagger
 springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/swagger-ui/
 springdoc.swagger-ui.operationsSorter=alpha
 springdoc.swagger-ui.tagsSorter=alpha
-
-#Profilo
-spring.profiles.active=dev
-
-
-#Connessione
-spring.datasource.url=jdbc:postgresql://localhost:5432/MemoFlowDB
-spring.datasource.username=postgres
-spring.datasource.password=Adriano5495
-
-#Configurazione
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
